# msvc-specification/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Swagger Microservices Bank
  description: Microservices Bank
  termsOfService: http://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: http://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://clientstore.swagger.io/api
paths:
...
  /products:
    get:
      tags:
        - PRODUCT
      summary: Returns all products from the system that the user has access to
      description: Returns all products from the system that the user has access to
      operationId: findProducts
      responses:
        '200':
          description: Product response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      tags:
        - PRODUCT
      summary: Creates a new product in the bank.  Duplicates are not allowed
      description: Creates a new product in the bank.  Duplicates are not allowed
      operationId: addProduct
      requestBody:
        description: Product to add to the store
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductDto'
      responses:
        '200':
          description: product response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /products/{id}:
    get:
      tags:
        - PRODUCT
      summary: Returns a product based on a single ID
      description: Returns a product based on a single ID
      operationId: find product by id
      parameters:
        - name: id
          in: path
          description: ID of product to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: product response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    put:
      tags:
        - PRODUCT
      summary: Update product
      description: This can only be done by the logged in product.
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          description: ID of product to fetch
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewProductDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewProductDto'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Product not found
        '405':
          description: Validation exception
    delete:
      tags:
        - PRODUCT
      summary: deletes a single product based on the ID supplied
      description: deletes a single product based on the ID supplied
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          description: ID of product to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: product deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
...
    ProductDto:
      allOf:
        - $ref: '#/components/schemas/NewProductDto'
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
    NewProductDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            name:
              type: "string"
            type:
              type: "string"
            category:
              type: "string"
            commission:
              type: "number"
              format: "double"
            limitDeposit:
              type: "number"
              format: "double"
            limitWithdrawal:
              type: "number"
              format: "double"
            line:
              type: "number"
              format: "double"
...
    ErrorDto:
      type: object
      required:
        - message
      properties:
        message:
          type: string